fix(menuBar component) fixes the logout functionality on the swipeable drawer child of menuBar
