+++ client/app.js
@@ -1,6 +1,4 @@
 import React from 'react'
-
-import { Navbar } from './components'
 import MenuAppBar from './components/menuBar'
 import Routes from './routes'
 
@@ -8,7 +6,6 @@ const App = () => {
   return (
     <div>
       <MenuAppBar />
-      <Navbar />
       <Routes />
     </div>
   )
